refactor: extraer DTOs de completed_service a archivo separado
- Crear completed_service.go con DTOs (Request, Response, converters)
- Seguir patrón existente (branch.go, diagnostic.go, evidence.go)
- Remover DTOs inline de completed_service_controller.go
